fix: optimizing how the dereferenced dosages are retrieved from the database;
Previously, the dereferenced dosages were using an optimized process
involving hundreds of HTTP requests resulting in a slow performance.

Now the dereferenced dosages are retrieved with a single HTTP request.

A a variable name was changed from "deref" to "stringified" to reflect the fact that the frontend no
longer performs the dereferencing, and is instead just stringifying the
dereferenced dosages.